Drop static actions usage in favor of more OO
1. Why is this change necessary?
As of now we're forced to use static actions to handle requests. Due to
that have lots of static methods, which make our code less testable.

Furthermore the way we have the code, we're already using singletons
by acknowleding that fact into the frameworks code, we're able to clean
our controllers from static methods.